sending messages tested and working
