+++ p2p/p2p_test.go
@@ -6,19 +6,22 @@ import (
 	crypto "github.com/libp2p/go-libp2p-crypto"
 )
 
-var testServiceConfig = &ServiceConfig{
+var testServiceConfigA = &ServiceConfig{
 	BootstrapNodes: []string{
-		"/ip4/104.131.131.82/tcp/4001/ipfs/QmaCpDMGvV2BGHeYERUEnRQAwe3N8SzbUtfsmvsqQLuvuJ",
-		"/ip4/104.236.179.241/tcp/4001/ipfs/QmSoLPppuBtQSGwKDZT2M73ULpjvfd3aZ6ha4oFGL1KrGM",
-		"/ip4/128.199.219.111/tcp/4001/ipfs/QmSoLSafTMBsPKadTEgaXctDQVcqN88CNLHXMkTNwMKPnu",
-		"/ip4/104.236.76.40/tcp/4001/ipfs/QmSoLV4Bbm51jM9C4gDYZQ9Cy3U6aXMJDAbzgu2fzaDs64",
-		"/ip4/178.62.158.247/tcp/4001/ipfs/QmSoLer265NRgSp2LA3dPaeykiS1J6DifTC88f5uVQKNAd",
+		GetLocalPeerInfo(),
 	},
 	Port: 7001,
 }
 
+var testServiceConfigB = &ServiceConfig{
+	BootstrapNodes: []string{
+		GetLocalPeerInfo(),
+	},
+	Port: 7002,
+}
+
 func TestBuildOpts(t *testing.T) {
-	_, err := testServiceConfig.buildOpts()
+	_, err := testServiceConfigA.buildOpts()
 	if err != nil {
 		t.Fatalf("TestBuildOpts error: %s", err)
 	}
@@ -50,7 +53,7 @@ func TestGenerateKey(t *testing.T) {
 }
 
 func TestStart(t *testing.T) {
-	s, err := NewService(testServiceConfig)
+	s, err := NewService(testServiceConfigA)
 	if err != nil {
 		t.Fatalf("NewService error: %s", err)
 	}
@@ -63,7 +66,7 @@ func TestStart(t *testing.T) {
 }
 
 func TestSend(t *testing.T) {
-	sa, err := NewService(testServiceConfig)
+	sa, err := NewService(testServiceConfigA)
 	if err != nil {
 		t.Fatalf("NewService error: %s", err)
 	}
@@ -73,7 +76,7 @@ func TestSend(t *testing.T) {
 		t.Errorf("Start error: %s", err)
 	}
 
-	sb, err := NewService(testServiceConfig)
+	sb, err := NewService(testServiceConfigB)
 	if err != nil {
 		t.Fatalf("NewService error: %s", err)
 	}
@@ -83,8 +86,12 @@ func TestSend(t *testing.T) {
 		t.Errorf("Start error: %s", err)
 	}
 
-	peer := sb.host.ID()
-	t.Log(peer.Pretty())
+	peer, err := sa.dht.FindPeer(sa.ctx,  sb.host.ID())
+	if err != nil {
+		t.Fatalf("could not find peer: %s", err)
+	}
+
+	t.Log(peer)
 	msg := []byte("hello there")
 	err = sa.Send(peer, msg)
 	if err != nil {
